requirements.txt
alembic==1.0.5
astroid==2.2.2
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
Flask-Login==0.4.1
Flask-Migrate==2.3.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.8
isort==4.3.10
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.1.0
mccabe==0.6.1
psycopg2==2.7.6.1
PyJWT==1.7.1
pylint==2.3.1
python-dateutil==2.7.5
python-editor==1.0.3
requests==2.21.0
six==1.11.0
SQLAlchemy==1.2.14
typed-ast==1.2.0
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.11.1
WTForms==2.2.1
